running the graphman as module; introducing a new entry point for cli
