+++ package.json
@@ -2,7 +2,7 @@
     "name": "@layer7/graphman",
     "version": "1.3.00",
     "description": "Layer7 API Gateway Graphman Client",
-    "main": "main.js",
+    "main": "modules/main.js",
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1"
     },
